feat #31 중간저장2
