+++ client/src/components/createStudy/TogglePublic.tsx
@@ -1,18 +1,10 @@
 import styled from "styled-components";
 import { useEffect, useState } from "react";
 
-interface ToggleProps {
-  form: any;
-  setForm: any;
-}
-
-const TogglePublic = ({ form, setForm }: ToggleProps) => {
+const TogglePublic = ({ onChange }: any) => {
   const [isTrue, setIsTrue] = useState(false);
   useEffect(() => {
-    setForm({
-      ...form,
-      openClose: isTrue,
-    });
+    onChange(isTrue);
   }, [isTrue]);
   return (
     <div className="toggleBox">
